requirements.txt
# REPLIC-AI project dependencies
# Install: .venv/bin/pip install -r requirements.txt

# PDF extraction
pymupdf>=1.26
pdfplumber>=0.11

# Data handling
pandas>=2.0
numpy>=1.24

# Statistics and ML
scikit-learn>=1.3
scipy>=1.10

# Visualization
matplotlib>=3.7
seaborn>=0.13

# Data access
requests>=2.31

# Image handling
Pillow>=10.0

# LLM APIs
google-genai>=1.0
python-dotenv>=1.0
